+++ src/Resources/app/storefront/src/main.js
@@ -1,5 +1,6 @@
 import CartFlyoutPlugin from "./plugin/cart-flyout/cart-flyout.plugin";
 import UspRotatorPlugin from "./plugin/usp-rotator/usp-rotator.plugin";
+import StickyHeaderNavPlugin from "./plugin/sticky-header/sticky-header.plugin";
 
 const PluginManager = window.PluginManager;
 
@@ -8,4 +9,9 @@ PluginManager.register(
     "UspRotator",
     UspRotatorPlugin,
     '[data-plugin="UspRotator"]'
+);
+PluginManager.register(
+    "StickyHeaderNav",
+    StickyHeaderNavPlugin,
+    '[data-sticky-header="true"]'
 );