+++ CHANGELOG.md
@@ -66,10 +66,18 @@
 -   `custom/plugins/StrixNLUxUpgrades/src/Resources/app/storefront/src/plugin/usp-rotator/usp-rotator.plugin.js`
 -   `custom/plugins/StrixNLUxUpgrades/src/Resources/app/storefront/src/main.js`
 -   `custom/plugins/StrixNLUxUpgrades/src/Resources/config/config.xml`
-### 2025-09-26 17:49 — 8bdb1904
+### 2025-09-30 19:14 — 4b212613
 - Added configurable USP bar with animation
 
 **Changed files:**
 
+- `custom/plugins/StrixNLUxUpgrades/README.md`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/app/storefront/dist/storefront/js/strix-n-l-ux-upgrades/strix-n-l-ux-upgrades.js`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/app/storefront/src/main.js`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/app/storefront/src/plugin/sticky-header/sticky-header.plugin.js`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/app/storefront/src/plugin/usp-rotator/usp-rotator.plugin.js`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/app/storefront/src/scss/base.scss`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/config/config.xml`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/views/storefront/layout/header/header.html.twig`
 - `custom/plugins/StrixNLUxUpgrades/src/Resources/views/storefront/layout/header/strix-usp.html.twig`